+++ app/src/main/java/com/kau/ttokttok/ui/component/auth/PasswordField.kt
@@ -1,0 +1,104 @@
+package com.kau.ttokttok.ui.component.auth
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.Lock
+import androidx.compose.material.icons.outlined.Visibility
+import androidx.compose.material.icons.outlined.VisibilityOff
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.LocalTextStyle
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun PasswordField(
+    value: String,
+    onValueChange: (String) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    var passwordVisible by rememberSaveable { mutableStateOf(false) }
+
+    Column(
+        modifier = modifier.fillMaxWidth()
+    ) {
+        Text(
+            text = "비밀번호",
+            color = Color.White,
+            style = MaterialTheme.typography.bodySmall,
+            modifier = Modifier.padding(bottom = 6.dp)
+        )
+
+        OutlinedTextField(
+            value = value,
+            onValueChange = onValueChange,
+            placeholder = {
+                Text(
+                    "비밀번호를 입력하세요",
+                    color = Color.White.copy(alpha = 0.6f) // placeholder:text-white/60
+                )
+            },
+            singleLine = true,
+            leadingIcon = {
+                Icon(
+                    imageVector = Icons.Outlined.Lock,
+                    contentDescription = null,
+                    tint = Color.White.copy(alpha = 0.6f),
+                    modifier = Modifier.size(18.dp)
+                )
+            },
+            trailingIcon = {
+                IconButton(onClick = { passwordVisible = !passwordVisible }) {
+                    Icon(
+                        imageVector = if (passwordVisible)
+                            Icons.Outlined.VisibilityOff
+                        else
+                            Icons.Outlined.Visibility,
+                        contentDescription = if (passwordVisible) "비밀번호 숨기기" else "비밀번호 보이기",
+                        tint = Color.White.copy(alpha = 0.6f)
+                    )
+                }
+            },
+            textStyle = LocalTextStyle.current.copy(color = Color.White),
+            visualTransformation = if (passwordVisible) VisualTransformation.None
+            else PasswordVisualTransformation(),
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(56.dp)
+                .clip(RoundedCornerShape(12.dp)), // rounded-xl
+            shape = RoundedCornerShape(12.dp),
+            colors = TextFieldDefaults.colors(
+                focusedContainerColor = Color.White.copy(alpha = 0.1f),   // bg-white/10
+                unfocusedContainerColor = Color.White.copy(alpha = 0.1f),
+                disabledContainerColor = Color.White.copy(alpha = 0.1f),
+
+                focusedIndicatorColor = Color.White.copy(alpha = 0.4f),   // focus:border-white/40
+                unfocusedIndicatorColor = Color.White.copy(alpha = 0.2f), // border-white/20
+
+                cursorColor = Color.White,
+                focusedLeadingIconColor = Color.White.copy(alpha = 0.6f),
+                unfocusedLeadingIconColor = Color.White.copy(alpha = 0.6f),
+                focusedTrailingIconColor = Color.White.copy(alpha = 0.6f),
+                unfocusedTrailingIconColor = Color.White.copy(alpha = 0.6f)
+            )
+        )
+    }
+}
