+++ app/src/main/java/com/kau/ttokttok/ui/component/auth/button/SignUpButton.kt
@@ -1,0 +1,34 @@
+package com.kau.ttokttok.ui.component.auth.button
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.OutlinedButton
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun SignupButton(
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    OutlinedButton(
+        onClick = onClick,
+        modifier = modifier
+            .fillMaxWidth()
+            .height(48.dp), // h-12
+        shape = RoundedCornerShape(12.dp), // rounded-xl
+        border = BorderStroke(1.dp, Color.White.copy(alpha = 0.3f)),
+        colors = ButtonDefaults.outlinedButtonColors(
+            containerColor = Color.Transparent,   // bg-transparent
+            contentColor = Color.White            // text-white
+        )
+    ) {
+        Text("회원가입 하기")
+    }
+}
