+++ app/src/main/java/com/kau/ttokttok/ui/component/common/Color.kt
@@ -1,0 +1,1 @@
+package com.kau.ttokttok.ui.component.common
