+++ app/src/main/java/com/kau/ttokttok/ui/compose/login/LoginScreen.kt
@@ -1,6 +1,5 @@
 package com.kau.ttokttok.ui.compose.login
 
-import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
@@ -8,10 +7,6 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.Email
-import androidx.compose.material.icons.outlined.Lock
-import androidx.compose.material.icons.outlined.Visibility
-import androidx.compose.material.icons.outlined.VisibilityOff
-import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.runtime.saveable.rememberSaveable
@@ -22,12 +17,15 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalFocusManager
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.kau.ttokttok.ui.component.auth.AuthInputField
+import com.kau.ttokttok.ui.component.auth.FrostedPanel
+import com.kau.ttokttok.ui.component.auth.PasswordField
+import com.kau.ttokttok.ui.component.auth.button.LoginButton
+import com.kau.ttokttok.ui.component.auth.button.SignupButton
 import com.kau.ttokttok.ui.component.common.background.StarField
 
 @Preview(showBackground = true)
@@ -35,19 +33,12 @@ import com.kau.ttokttok.ui.component.common.background.StarField
 fun LoginScreen(
     modifier: Modifier = Modifier,
     onClickLogin: (email: String, password: String) -> Unit = { _, _ -> },
-    onClickRegister: () -> Unit = {}
+    onClickRegister: () -> Unit = {},
+    onRequestTempPassword: () -> Unit = {}
 ) {
-    val focus = LocalFocusManager.current
-    val scroll = rememberScrollState()
-
-    // 화면 상태 (디자인/미리보기용)
     var email by rememberSaveable { mutableStateOf("") }
-    var pw by rememberSaveable { mutableStateOf("") }
-    var pwVisible by rememberSaveable { mutableStateOf(false) }
-    var showCard by remember { mutableStateOf(false) }
-
-    // 간단한 등장 애니메이션
-    LaunchedEffect(Unit) { showCard = true }
+    var password by rememberSaveable { mutableStateOf("") }
+    var passwordVisible by rememberSaveable { mutableStateOf(false) }
 
     Box(
         modifier = modifier
@@ -62,17 +53,17 @@ fun LoginScreen(
                     end = Offset.Infinite
                 )
             )
+            .statusBarsPadding()
     ) {
-        // 별빛 배경
         StarField(
-            modifier = Modifier
+            modifier = modifier
                 .fillMaxSize()
                 .alpha(0.35f),
         )
 
         // 페이드 보카시
         Box(
-            Modifier
+            modifier
                 .matchParentSize()
                 .background(
                     Brush.radialGradient(
@@ -84,7 +75,7 @@ fun LoginScreen(
         )
 
         Column(
-            modifier = Modifier
+            modifier = modifier
                 .fillMaxSize()
                 .verticalScroll(rememberScrollState())
                 .padding(horizontal = 20.dp),
@@ -92,50 +83,49 @@ fun LoginScreen(
             verticalArrangement = Arrangement.Center
         ) {
             FrostedPanel(
-                modifier = Modifier
+                modifier = modifier
                     .fillMaxWidth()
-                    .padding(vertical = 32.dp)
                     .padding(horizontal = 24.dp)
             ) {
                 // 로고/타이틀 영역
                 Column(
-                    modifier = Modifier
+                    modifier = modifier
                         .fillMaxWidth()
                         .padding(bottom = 24.dp),
                     horizontalAlignment = Alignment.CenterHorizontally
                 ) {
                     // 애니메이션 들어가는 로고
                     Box(
-                        modifier = Modifier
+                        modifier = modifier
                             .size(64.dp) // w-16 h-16
                             .clip(RoundedCornerShape(16.dp)) // rounded-2xl
-                            .background(Color.White.copy(alpha = 0.2f)) // bg-white/20
+                            .background(Color.White.copy(alpha = 0.2f))
                             .padding(12.dp), // 내부 여백
                         contentAlignment = Alignment.Center
                     ) {
                         Box(
-                            modifier = Modifier
+                            modifier = modifier
                                 .size(32.dp) // w-8 h-8
                                 .clip(RoundedCornerShape(8.dp))
                                 .background(Color.White),
                             contentAlignment = Alignment.Center
                         ) {
                             Box(
-                                modifier = Modifier
+                                modifier = modifier
                                     .size(16.dp) // w-4 h-4
                                     .clip(RoundedCornerShape(2.dp))
-                                    .background(Color(0xFF2563EB)) // bg-blue-600
+                                    .background(Color(0xFF2563EB))
                             )
                         }
                     }
 
-                    Spacer(Modifier.height(16.dp))
+                    Spacer(modifier.height(16.dp))
 
                     Text(
                         text = "똑똑",
                         style = MaterialTheme.typography.headlineSmall,
                         color = Color.White,
-                        modifier = Modifier.padding(bottom = 4.dp)
+                        modifier = modifier.padding(bottom = 4.dp)
                     )
 
                     Text(
@@ -146,237 +136,65 @@ fun LoginScreen(
                     )
                 }
 
-                EmailField(
+                AuthInputField(
+                    onValueChange = { email = it },
                     value = email,
-                    onValueChange = { email = it }
+                    headerText = "이메일",
+                    imageVector = Icons.Outlined.Email,
+                    exampleText = "example@gmail.com"
                 )
 
-                Spacer(Modifier.height(16.dp))
+                Spacer(modifier.height(16.dp))
 
                 PasswordField(
-                    value = pw,
-                    onValueChange = { pw= it }
+                    value = password,
+                    onValueChange = { password= it }
                 )
 
-                Spacer(Modifier.height(32.dp))
+                Spacer(modifier.height(32.dp))
 
                 LoginButton(
                     onClick = {
-                        onClickLogin(email, pw)
+                        onClickLogin(email, password)
                     }
                 )
 
-                Spacer(Modifier.height(32.dp))
+                Spacer(modifier.height(32.dp))
 
                 SignupButton(
                     onClick = onClickRegister
                 )
             }
-        }
-    }
-}
 
-@Composable
-fun FrostedPanel(
-    modifier: Modifier = Modifier,
-    // rounded-3xl ≈ 24.dp
-    corner: RoundedCornerShape = RoundedCornerShape(24.dp),
-    content: @Composable ColumnScope.() -> Unit
-) {
-    val panelColor = Color.White.copy(alpha = 0.10f)   // bg-white/10
-    val borderColor = Color.White.copy(alpha = 0.20f)  // border-white/20
-
-    // shadow-2xl 느낌 → shadowElevation 크게
-    Surface(
-        shape = corner,
-        color = panelColor,
-        tonalElevation = 0.dp,
-        shadowElevation = 0.dp,
-        border = BorderStroke(1.dp, borderColor),
-        modifier = modifier
-            .clip(corner)
-    ) {
-        Column(Modifier.padding(24.dp)) { // p-8
-            content()
-        }
-    }
-}
-
-@Composable
-fun EmailField(
-    value: String,
-    onValueChange: (String) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    var isFocused by remember { mutableStateOf(false) }
-
-    Column(
-        modifier = modifier.fillMaxWidth()
-    ) {
-        Text(
-            text = "이메일",
-            color = Color.White,
-            style = MaterialTheme.typography.bodySmall,
-            modifier = Modifier.padding(bottom = 6.dp)
-        )
-
-        OutlinedTextField(
-            value = value,
-            onValueChange = onValueChange,
-            placeholder = {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(top = 12.dp),
+                horizontalArrangement = Arrangement.Center,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
                 Text(
-                    "example@gmail.com",
-                    color = Color.White.copy(alpha = 0.6f) // placeholder:text-white/60
+                    text = "비밀번호를 잊으셨나요?",
+                    style = MaterialTheme.typography.bodySmall,
+                    color = Color.White.copy(alpha = 0.7f),
+                    modifier = Modifier.padding(end = 4.dp)
                 )
-            },
-            singleLine = true,
-            leadingIcon = {
-                Icon(
-                    imageVector = Icons.Outlined.Email,
-                    contentDescription = null,
-                    tint = Color.White.copy(alpha = 0.6f), // 아이콘 색
-                    modifier = Modifier.size(18.dp)
-                )
-            },
-            textStyle = LocalTextStyle.current.copy(color = Color.White), // 입력 텍스트 색
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(56.dp)
-                .clip(RoundedCornerShape(12.dp)), // rounded-xl
-            shape = RoundedCornerShape(12.dp),
-            colors = TextFieldDefaults.colors(
-                focusedContainerColor = Color.White.copy(alpha = 0.1f),   // bg-white/10
-                unfocusedContainerColor = Color.White.copy(alpha = 0.1f), // bg-white/10
-                disabledContainerColor = Color.White.copy(alpha = 0.1f),
 
-                focusedIndicatorColor = Color.White.copy(alpha = 0.4f),   // focus:border-white/40
-                unfocusedIndicatorColor = Color.White.copy(alpha = 0.2f), // border-white/20
-
-                cursorColor = Color.White,
-                focusedLeadingIconColor = Color.White.copy(alpha = 0.6f),
-                unfocusedLeadingIconColor = Color.White.copy(alpha = 0.6f),
-                focusedTrailingIconColor = Color.White.copy(alpha = 0.6f),
-                unfocusedTrailingIconColor = Color.White.copy(alpha = 0.6f)
-            )
-        )
-    }
-}
-
-@Composable
-fun PasswordField(
-    value: String,
-    onValueChange: (String) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    var passwordVisible by rememberSaveable { mutableStateOf(false) }
-
-    Column(
-        modifier = modifier.fillMaxWidth()
-    ) {
-        Text(
-            text = "비밀번호",
-            color = Color.White,
-            style = MaterialTheme.typography.bodySmall,
-            modifier = Modifier.padding(bottom = 6.dp)
-        )
-
-        OutlinedTextField(
-            value = value,
-            onValueChange = onValueChange,
-            placeholder = {
-                Text(
-                    "비밀번호를 입력하세요",
-                    color = Color.White.copy(alpha = 0.6f) // placeholder:text-white/60
-                )
-            },
-            singleLine = true,
-            leadingIcon = {
-                Icon(
-                    imageVector = Icons.Outlined.Lock,
-                    contentDescription = null,
-                    tint = Color.White.copy(alpha = 0.6f),
-                    modifier = Modifier.size(18.dp)
-                )
-            },
-            trailingIcon = {
-                IconButton(onClick = { passwordVisible = !passwordVisible }) {
-                    Icon(
-                        imageVector = if (passwordVisible)
-                            Icons.Outlined.VisibilityOff
-                        else
-                            Icons.Outlined.Visibility,
-                        contentDescription = if (passwordVisible) "비밀번호 숨기기" else "비밀번호 보이기",
-                        tint = Color.White.copy(alpha = 0.6f)
+                TextButton(
+                    onClick = onRequestTempPassword,
+                    modifier = Modifier.wrapContentWidth(),
+                    colors = ButtonDefaults.textButtonColors(
+                        contentColor = Color(0xFF64B5F6)
+                    )
+                ) {
+                    Text(
+                        text = "임시 비밀번호 발급하기",
+                        style = MaterialTheme.typography.bodySmall,
+                        fontWeight = FontWeight.Medium,
+                        textAlign = TextAlign.Center
                     )
                 }
-            },
-            textStyle = LocalTextStyle.current.copy(color = Color.White),
-            visualTransformation = if (passwordVisible) VisualTransformation.None
-            else PasswordVisualTransformation(),
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(56.dp)
-                .clip(RoundedCornerShape(12.dp)), // rounded-xl
-            shape = RoundedCornerShape(12.dp),
-            colors = TextFieldDefaults.colors(
-                focusedContainerColor = Color.White.copy(alpha = 0.1f),   // bg-white/10
-                unfocusedContainerColor = Color.White.copy(alpha = 0.1f),
-                disabledContainerColor = Color.White.copy(alpha = 0.1f),
-
-                focusedIndicatorColor = Color.White.copy(alpha = 0.4f),   // focus:border-white/40
-                unfocusedIndicatorColor = Color.White.copy(alpha = 0.2f), // border-white/20
-
-                cursorColor = Color.White,
-                focusedLeadingIconColor = Color.White.copy(alpha = 0.6f),
-                unfocusedLeadingIconColor = Color.White.copy(alpha = 0.6f),
-                focusedTrailingIconColor = Color.White.copy(alpha = 0.6f),
-                unfocusedTrailingIconColor = Color.White.copy(alpha = 0.6f)
-            )
-        )
-    }
-}
-
-@Composable
-fun LoginButton(
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Button(
-        onClick = onClick,
-        modifier = modifier
-            .fillMaxWidth()
-            .height(48.dp), // h-12
-        shape = RoundedCornerShape(12.dp), // rounded-xl
-        colors = ButtonDefaults.buttonColors(
-            containerColor = Color.White.copy(alpha = 0.2f),   // bg-white/20
-            contentColor = Color.White                        // text-white
-        ),
-        contentPadding = PaddingValues(0.dp) // Tailwind 기본처럼 꽉 채운 텍스트
-    ) {
-        Text(
-            "로그인",
-            style = MaterialTheme.typography.bodyLarge
-        )
-    }
-}
-
-@Composable
-fun SignupButton(
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    OutlinedButton(
-        onClick = onClick,
-        modifier = modifier
-            .fillMaxWidth()
-            .height(48.dp), // h-12
-        shape = RoundedCornerShape(12.dp), // rounded-xl
-        border = BorderStroke(1.dp, Color.White.copy(alpha = 0.3f)),
-        colors = ButtonDefaults.outlinedButtonColors(
-            containerColor = Color.Transparent,   // bg-transparent
-            contentColor = Color.White            // text-white
-        )
-    ) {
-        Text("회원가입 하기")
+            }
+        }
     }
 }
